# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.2.0",
    "title": "FOUNDLUNTEER API",
    "description":"v1.1.0\n-Add Show saved job for individual\n-Create universal Add, Login, And Get For User\n\nv1.2.0\n-Add get and post image profile for user\n-Register API for Individual\n-Get detail job for Organization"
  },
  "host": "aws.senna-annaba.my.id",
  "basePath": "/",
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["https"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter the token with the Bearer: bearer [token]"
    }
  },
  "paths": {
    "/user/add": {
      "post": {
        "tags": ["User"],
        "summary": "Add User",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "User",
          "description": "Add New User",
          "schema": {
            "type": "object",
            "required": [
              "email",
              "password",
              "name",
              "address",
              "phone",
              "role"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["User"],
        "summary": "Login For User",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "User",
          "description": "Login For User",
          "schema": {
            "type": "object",
            "required": [
              "email",
              "password"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/get": {
      "get": {
        "tags": ["User"],
        "summary": "Get User Profile",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/image":{
      "post": {
        "tags": ["User"],
        "summary": "Add Individual or Organization Image Profile",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "parameters": [{
          "in": "formData",
          "name": "image",
          "description": "jpg/jpeg/png",
          "type": "file"
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/getimage": {
      "get": {
        "tags": ["User"],
        "summary": "Get User Image",
        "produces": ["image/png"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          }
        }
      }
    },
...
    "/organization/update": {
      "post": {
        "tags": ["Organization"],
        "summary": "Update For Organization",
        "description": "All body request optional",
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Individual",
          "description": "Add New Individual",
          "schema": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "leader": {
                "type": "string"
              },
              "description": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/organization/getjob": {
      "get": {
        "tags": ["Organization"],
        "summary": "Get Organization Job",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/organization/jobdetail": {
      "post": {
        "tags": ["Organization"],
        "summary": "Get Organization Job Detail With Registrant",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "parameters": [{
          "in": "body",
          "name": "Job Id",
          "description": "Get Detail Job And Registrant",
          "schema": {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/organization/add": {
      "post": {
        "deprecated": true,
        "tags": ["Organization"],
        "summary": "Add Organization",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Organization",
          "description": "Add New Organization",
          "schema": {
            "type": "object",
            "required": [
              "email",
              "password",
              "name",
              "address",
              "phone",
              "leader"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "leader": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/organization/login": {
      "post": {
        "deprecated": true,
        "tags": ["Organization"],
        "summary": "Login For Organization",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "Organization",
          "description": "Login For Organization",
          "schema": {
            "type": "object",
            "required": [
              "email",
              "password"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/organization/get": {
      "get": {
        "deprecated": true,
        "tags": ["Organization"],
        "summary": "Get Organization Profile",
        "produces": ["application/json"],
        "security": [
          {
            "bearerAuth":[]
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Try it out for response"

          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}